+++ Source/Public/memory/allocators.h
@@ -3,7 +3,7 @@
 namespace edvar::memory {
 template <typename storage_type> class heap_allocator {
 public:
-    storage_type* allocate(size_t count) { return static_cast<storage_type*>(::operator new[](sizeof(storage_type) * count)); }
+    storage_type* allocate(uint32 count) { return static_cast<storage_type*>(::operator new[](sizeof(storage_type) * count)); }
     void deallocate(storage_type* ptr) { ::operator delete[](ptr); }
 };
 } // namespace edvar::memory
